+++ R/class_options.R
@@ -317,7 +317,7 @@ options_class <- R6::R6Class(
       self$memory %|||% "persistent"
     },
     get_garbage_collection = function() {
-      self$garbage_collection %|||% FALSE
+      self$garbage_collection %|||% 0L
     },
     get_deployment = function() {
       self$deployment %|||% "worker"
@@ -407,7 +407,7 @@ options_class <- R6::R6Class(
     },
     set_garbage_collection = function(garbage_collection) {
       self$validate_garbage_collection(garbage_collection)
-      self$garbage_collection <- garbage_collection
+      self$garbage_collection <- as.integer(garbage_collection)
     },
     set_deployment = function(deployment) {
       self$validate_deployment(deployment)
@@ -523,8 +523,10 @@ options_class <- R6::R6Class(
       tar_assert_flag(memory, c("persistent", "transient"))
     },
     validate_garbage_collection = function(garbage_collection) {
-      tar_assert_lgl(garbage_collection)
+      tar_assert_dbl(garbage_collection)
       tar_assert_scalar(garbage_collection)
+      tar_assert_none_na(garbage_collection)
+      tar_assert_ge(garbage_collection, 0L)
     },
     validate_deployment = function(deployment) {
       tar_assert_flag(deployment, c("worker", "main"))