+++ R/tar_make_future.R
@@ -20,10 +20,6 @@
 #' @inheritParams tar_make
 #' @param workers Positive integer, maximum number of transient
 #'   `future` workers allowed to run at any given time.
-#' @param garbage_collection Logical of length 1, whether to run garbage
-#'   collection on the main process before sending a target to a worker.
-#'   Independent from the `garbage_collection` argument of [tar_target()],
-#'   which controls garbage collection on the worker.
 #' @examples
 #' if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
 #' tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
@@ -53,7 +49,7 @@ tar_make_future <- function(
   envir = parent.frame(),
   script = targets::tar_config_get("script"),
   store = targets::tar_config_get("store"),
-  garbage_collection = targets::tar_config_get("garbage_collection")
+  garbage_collection = NULL
 ) {
   tar_assert_allow_meta("tar_make_future", store)
   force(envir)
@@ -79,9 +75,16 @@ tar_make_future <- function(
   tar_assert_none_na(seconds_reporter)
   tar_assert_ge(seconds_reporter, 0)
   tar_deprecate_seconds_interval(seconds_interval)
-  tar_assert_lgl(garbage_collection)
-  tar_assert_scalar(garbage_collection)
-  tar_assert_none_na(garbage_collection)
+  if_any(
+    is.null(garbage_collection),
+    NULL,
+    tar_warn_deprecate(
+      "The garbage_collection argument of tar_make_future() was deprecated ",
+      "in version 1.8.0.9004 (2024-10-22). The garbage_collection ",
+      "argument of tar_option_set() is more unified and featureful now. ",
+      "Please have a look at its documentation."
+    )
+  )
   targets_arguments <- list(
     path_store = store,
     names_quosure = rlang::enquo(names),
@@ -90,7 +93,6 @@ tar_make_future <- function(
     seconds_meta_append = seconds_meta_append,
     seconds_meta_upload = seconds_meta_upload,
     seconds_reporter = seconds_reporter,
-    garbage_collection = garbage_collection,
     workers = workers
   )
   out <- callr_outer(
@@ -115,7 +117,6 @@ tar_make_future_inner <- function(
   seconds_meta_append,
   seconds_meta_upload,
   seconds_reporter,
-  garbage_collection,
   workers
 ) {
   names <- tar_tidyselect_eval(names_quosure, pipeline_get_names(pipeline))
@@ -129,7 +130,6 @@ tar_make_future_inner <- function(
     seconds_meta_append = seconds_meta_append,
     seconds_meta_upload = seconds_meta_upload,
     seconds_reporter = seconds_reporter,
-    garbage_collection = garbage_collection,
     workers = workers
   )$run()
   invisible()